Project Filmorate v0.2.2
refactor: перенести логирование из методов createFilm() и updateFilm() класса FilmService в методы createFilm() и updateFilm() класса FilmController;
refactor: перенести логирование из методов createUser() и updateUser() класса UserService в методы createUser() и updateUser() класса UserController;
feat: добавить метод getUsersIds() в интерфейс UserStorage;
fix: заменить в методе addLikeToFilm() класса FilmService условие проверки в первом блоке If на проверку по Id пользователя;
fix: заменить в методе removeLikeOfFilm() класса FilmService условие проверки в первом блоке If на проверку по Id пользователя;
fix: заменить в методе addFriend() класса UserService условие проверки в первом блоке If на проверку по Id друга;
fix: заменить в методе removeFriend() класса UserService условие проверки в первом блоке If на проверку по Id друга;
fix: заменить тип у переменных, ответственных за хранение ID, с int/Integer на long/Long;

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -11,7 +11,7 @@ import java.util.Set;
 @Data
 @AllArgsConstructor
 public class Film {
-    private int id;
+    private long id;
     @NonNull
     private String name;
     @NonNull
@@ -19,5 +19,5 @@ public class Film {
     @NonNull
     private LocalDate releaseDate;
     private int duration;
-    private final Set<Integer> likes = new HashSet<>();
+    private final Set<Long> likes = new HashSet<>();
 }
